Refactor ledger entry imports and create ledger entry class
- Updated import paths for LedgerEntry and LedgerType from `ledger_entry` to `ledger_entry_class` across multiple files.
- Introduced a new `ledger_entry_class.py` file containing the `LedgerEntry` class and related exceptions.
- Added a new `ledger_type_class.py` file defining the `LedgerType` enumeration for accounting transactions.
- Ensured all references to ledger entries in the codebase are consistent with the new structure.
