+++ src/v4vapp_backend_v2/actions/payment_success.py
@@ -8,7 +8,7 @@ from v4vapp_backend_v2.accounting.ledger_account_classes import (
     LiabilityAccount,
     RevenueAccount,
 )
-from v4vapp_backend_v2.accounting.ledger_entry import LedgerEntry, LedgerType
+from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry, LedgerType
 from v4vapp_backend_v2.actions.hive_to_lnd import (
     calculate_hive_return_change,
     lightning_payment_sent,
