+++ tests/actions/pause_test_hive_to_lightning_live_lightning_node.py
@@ -14,7 +14,7 @@ from v4vapp_backend_v2.accounting.balance_sheet import (
     balance_sheet_printout,
     generate_balance_sheet_pandas_from_accounts,
 )
-from v4vapp_backend_v2.accounting.ledger_entry import LedgerEntry, get_ledger_entry
+from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry, get_ledger_entry
 from v4vapp_backend_v2.actions.hive_to_lnd import (
     HiveToLightningError,
     process_hive_to_lightning,
