+++ tests/actions/pause_test_process_tracked_events.py
@@ -15,7 +15,7 @@ import pytest
 from bson import json_util
 from nectar.hive import Hive
 
-from v4vapp_backend_v2.accounting.ledger_entry import LedgerEntry, LedgerEntryDuplicateException
+from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry, LedgerEntryDuplicateException
 from v4vapp_backend_v2.actions.hive_to_lnd import (
     HiveToLightningError,
     complete_hive_to_lightning,
