+++ tests/actions/test_full_stack.py
@@ -16,7 +16,7 @@ from v4vapp_backend_v2.accounting.account_balances import (
     get_keepsats_balance,
     keepsats_balance_printout,
 )
-from v4vapp_backend_v2.accounting.ledger_entry import LedgerEntry, LedgerType
+from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry, LedgerType
 from v4vapp_backend_v2.config.setup import HiveRoles, InternalConfig, logger
 from v4vapp_backend_v2.database.db_pymongo import DBConn
 from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConversion
